move useParticipantInfo hook
